Avance backend de Horario  ( conexcion con DB)
NullPointer Exception :astonished:
