+++ src/main/java/edu/eci/cvds/entities/Recurso.java
@@ -18,6 +18,7 @@ public class Recurso {
     
     
     
+    
 
 
     public Recurso(int id,String nombre, String ubicacion, String tipo, int capacidad, String estado, java.sql.Time horaInicio, java.sql.Time horaFin) {
