+++ src/main/java/edu/eci/cvds/managedbeans/ComunidadBean.java
@@ -87,97 +87,6 @@ public class ComunidadBean extends BasePageBean {
         this.recurrenciaSeleccionada = recurrenciaSeleccionada;
     }
 
-    public void onEventSelect(SelectEvent selectEvent) {
-        event = (ScheduleEvent) selectEvent.getObject();
-    }
-
-    public void onDateSelect(SelectEvent selectEvent) {
-        event = new DefaultScheduleEvent("", (Date) selectEvent.getObject(), (Date) selectEvent.getObject());
-    }
-
-    public void onEventMove(ScheduleEntryMoveEvent event) {
-        FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO, "Event moved", "Day delta:" + event.getDayDelta() + ", Minute delta:" + event.getMinuteDelta());
-
-        addMessage(message);
-    }
-
-    public void onEventResize(ScheduleEntryResizeEvent event) {
-        FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO, "Event resized", "Day delta:" + event.getDayDelta() + ", Minute delta:" + event.getMinuteDelta());
-
-        addMessage(message);
-    }
-
-    private void addMessage(FacesMessage message) {
-        FacesContext.getCurrentInstance().addMessage(null, message);
-    }
-
-    public void deleteEvent() {
-        eventModel.deleteEvent(event);
-    }
-
-    public ScheduleModel getEventModel() {
-        return eventModel;
-    }
-
-    public void setEventModel(ScheduleModel eventModel) {
-        this.eventModel = eventModel;
-    }
-
-    public ScheduleEvent getEvent() {
-        return event;
-    }
-
-    public void setEvent(ScheduleEvent event) {
-        this.event = event;
-    }
-
-    public void addEvent() {
-        if (event.getId() == null) {
-            /* En esta linea se deben llamar al metodo recursion,con los parametros especificados
-            es decir addEvent tambien deberia tener tipoReserva y Duracion todo depende de los campos 
-            que los de vista deben hacer :V
-            
-            recursion();
-            */
-        } else {
-            eventModel.updateEvent(event);
-        }
-
-        event = new DefaultScheduleEvent();
-    }
-
-    public void recursion(TipoReserva res,int duracion) {
-        eventModel.addEvent(event);
-            for (int i = 1; i < duracion; i++) {
-                
-                event = new DefaultScheduleEvent(event.getTitle(), sumaFecha(event.getStartDate(),res), sumaFecha(event.getEndDate(), res));
-                System.out.println(event.getStartDate());
-                eventModel.addEvent(event);
-            }
-      
-       
-
-    }
-
-    private Date sumaFecha(Date fecha, TipoReserva res) {
-        dateFormat.format(fecha);
-        Calendar c = Calendar.getInstance();
-        c.setTime(fecha);
-
-        switch (res) {
-            case Diario:
-                c.add(Calendar.DATE, 1);
-                break;
-            case Semanal:
-                c.add(Calendar.DATE, 7);
-                break;
-            case Mensual:
-                c.add(Calendar.MONTH, 1);
-                break;
-        }
-        Date result = c.getTime();
-        dateFormat.format(result);
-        return result;
-    }
+    
 
 }
