+++ src/main/java/edu/eci/cvds/persistence/mybatisimpl/mappers/ReservaMapper.java
@@ -19,6 +19,6 @@ public interface ReservaMapper {
 
     public void salvar(@Param("reserva") Reserva rev);
 
-    public void remover(@Param("reserva") Reserva rev);
+    public void remover(@Param("reservaId") int rev);
 
 }
